requirements.txt
openai
langchain
chromadb
pypdf
python-dotenv
streamlit
sentence-transformers
langchain_community
langchain_huggingface
langchain
huggingface_hub
transformers
langchain_chroma
hf_xet
google-generativeai>=0.3.0
# Additional requirements for enhanced PDF processing
pymupdf>=1.22.3
pytesseract>=0.3.10
pdf2image>=1.16.3
poppler-utils